requirements.txt
click~=8.0.3
pillow~=8.4.0
pandas~=1.3.4
numpy~=1.21.5
scipy~=1.7.3
seaborn~=0.11.2
h5py~=3.6.0
tqdm~=4.62.3
sklearn~=0.0
scikit-learn~=1.0.2
dtw~=1.4.0